biginteger

+++ usersys/models.py
@@ -313,7 +313,7 @@ class userAttachments(MyModel):
 
 class userEvents(MyModel):
     user = MyForeignKey(MyUser,related_name='user_userEvents', blank=True, on_delete=CASCADE)
-    com_id = models.IntegerField(blank=True, null=True, help_text='全库项目id')
+    com_id = models.BigIntegerField(blank=True, null=True, help_text='全库项目id')
     comshortname = models.CharField(max_length=64, blank=True, null=True)
     investDate = models.DateTimeField(blank=True, null=True)
     deleteduser = MyForeignKey(MyUser,blank=True, null=True,related_name='userdelete_userEvents')
